+++ src/main/java/com/springboot/jewellerysystem/controller/CompanyDetailController.java
@@ -14,9 +14,12 @@ import org.springframework.web.multipart.MultipartFile;
 import com.springboot.jewellerysystem.entity.CompanyDetail;
 import com.springboot.jewellerysystem.service.CompanyDetailService;
 import com.springboot.jewellerysystem.util.FileUploadUtil;
+import com.springboot.jewellerysystem.util.Helper;
 
 import java.io.IOException;
 import java.util.List;
+
+import javax.servlet.http.HttpSession;
 
 @Controller
 @RequestMapping(value = "admin/companyDetail")
@@ -29,6 +32,8 @@ public class CompanyDetailController {
 
 	@GetMapping(value = "/index")
 	public String companyDetails(Model model, @RequestParam(name = "keyword", defaultValue = "") String keyword) {
+		if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
 		List<CompanyDetail> companyDetails = companyDetailService.getAllCompanyDetail();
 		model.addAttribute("listCompanyDetails", companyDetails);
 		model.addAttribute("keyword", keyword);
@@ -37,19 +42,26 @@ public class CompanyDetailController {
 
 	@GetMapping(value = "/create")
 	public String formCompanyDetails(Model model) {
+		if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
 		model.addAttribute("companyDetail", new CompanyDetail());
 		return "admin/entry/companyDetail_entry";
 	}
 
 	@GetMapping(value = "/delete/{id}")
-	public String deleteCompanyDetail(@PathVariable(value = "id") Integer id, String keyword) {
+	public String deleteCompanyDetail(@PathVariable(value = "id") Integer id, String keyword, HttpSession session) {
+		if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
 		companyDetailService.removeCompanyDetail(id);
+		session.setAttribute("msg", "deleted");
 		return "redirect:/admin/companyDetail/index?keyword=" + keyword;
 	}
 
 	@GetMapping(value = "/update/{id}")
 	public String updateCompanyDetail(@PathVariable(value = "id") Integer id, Model model) {
-		CompanyDetail companyDetail = companyDetailService.loadCompanyDetailById(id);
+		
+		if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}CompanyDetail companyDetail = companyDetailService.loadCompanyDetailById(id);
 		model.addAttribute("companyDetail", companyDetail);
 		List<CompanyDetail> companyDetails = companyDetailService.getAllCompanyDetail();
 		model.addAttribute("listCompanyDetails", companyDetails);
@@ -57,8 +69,11 @@ public class CompanyDetailController {
 	}
 
 	@PostMapping(value = "/save")
-	public String save(CompanyDetail companyDetail, @RequestParam("file") MultipartFile file) throws IOException {
-
+	public String save(CompanyDetail companyDetail, @RequestParam("file") MultipartFile file, HttpSession session) throws IOException {
+		String msg = "inserted";
+		if(companyDetail.getId() != 0) {
+			msg = "updated";
+		}
 		String fileName = StringUtils.cleanPath(file.getOriginalFilename());
 		if (fileName.length() > 3) {
 		companyDetail.setLogo(fileName);
@@ -66,7 +81,12 @@ public class CompanyDetailController {
 		FileUploadUtil.saveFile(uploadDir, fileName, file);
 		}
 		
-		companyDetailService.createOrUpdateCompanyDetail(companyDetail);
+		CompanyDetail c = companyDetailService.createOrUpdateCompanyDetail(companyDetail);
+		if(c != null) {
+			session.setAttribute("msg", "inserted");
+		}else {
+			session.setAttribute("error","error");
+		}
 		return "redirect:/admin/companyDetail/index";
 	}
